Authorization Implemented

+++ src/api/notes/handler.js
@@ -12,9 +12,12 @@ class NotesHandler {
 
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
-    const { title = 'Untitled', body, tags } = request.payload;
+    const { title = 'untitled', body, tags } = request.payload;
+    const { id: credentialId } = request.auth.credentials;
 
-    const noteId = await this._service.addNote({ title, body, tags });
+    const noteId = await this._service.addNote({
+      title, body, tags, owner: credentialId,
+    });
 
     const response = h.response({
       status: 'success',
@@ -27,9 +30,9 @@ class NotesHandler {
     return response;
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
-
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -40,9 +43,10 @@ class NotesHandler {
 
   async getNoteByIdHandler(request) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
 
+    await this._service.verifyNoteOwner(id, credentialId);
     const note = await this._service.getNoteById(id);
-
     return {
       status: 'success',
       data: {
@@ -51,22 +55,26 @@ class NotesHandler {
     };
   }
 
-  async putNoteByIdHandler(request, h) {
+  async putNoteByIdHandler(request) {
     this._validator.validateNotePayload(request.payload);
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
 
-    await this._service.editNoteById(id, request.payload);
+    await this._service.verifyNoteOwner(id, credentialId);
+    this._service.editNoteById(id, request.payload);
 
-    return h.response({
+    return {
       status: 'success',
       message: 'Catatan berhasil diperbarui',
-    });
+    };
   }
 
   async deleteNoteByIdHandler(request) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
 
-    await this._service.deleteNoteById(id);
+    await this._service.verifyNoteOwner(id, credentialId);
+    this._service.deleteNoteById(id);
 
     return {
       status: 'success',
